+++ src/kettled/daemon/__init__.py
@@ -1,0 +1,9 @@
+from .pipes import pipe_command, read_pipe
+from .scheduler import Scheduler
+from .daemon import get_daemon_pid, Daemon
+
+__all__ = [
+    "pipe_command",
+    "read_pipe",
+    "Scheduler"
+]
