+++ src/kettled/handlers/import_handler.py
@@ -2,8 +2,7 @@ from sys import stderr
 from json import dumps
 from datetime import datetime
 from typing import Callable, Union
-from kettled.daemon.daemon import get_daemon_pid
-from kettled.daemon.pipes import pipe_command
+from kettled.daemon import get_daemon_pid, pipe_command
 from kettled.constants import (
     COMMANDS_ENUM, ERROR_MESSAGES_ENUM, MESSAGES_ENUM,
     UPDATE_EVENT_PARAMETERS_ENUM, PIPE_COMMANDS_ENUM,
