+++ tests/scheduler/test_scheduler_set.py
@@ -1,6 +1,6 @@
 import pytest
 from datetime import datetime
-from kettled.daemon.scheduler import Scheduler
+from kettled.daemon import Scheduler
 from kettled.constants import ERROR_MESSAGES_ENUM, EVENT_PARAMETERS_ENUM
 from tests.utils import get_future_datetime
 import tests.seeds as seeds
